+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changes for lovely.esdb
 unreleased
 ==========
 
+ - do not apply None for ObjectProperty if property not found
  - fixed document.get_source for ObjectProperty handling
 
 2016/07/27 0.3.0